==> prad_normal_vs_tumor/04_vulcano_plot.Rmd <==
---
title: "Vulcano plot new"
author: "Laura Lange"
date: '2022-05-25'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load data

```{r}
setwd('..')
tcga_tumor_norm <- readRDS("data/PRAD_TvsN.rds")
```
Load data
```{r}
# Fuse data frame PRAD normal with PRAD tumor
PRAD_tumor_normal <- cbind(tcga_tumor_norm[["tumor"]], tcga_tumor_norm[["normal"]])

class(PRAD_tumor_normal)
```

Calculate mean for each gene 

```{r}
mean_tumor <- apply(PRAD_tumor_normal[ ,1:52], 1, mean)

mean_normal <- apply(PRAD_tumor_normal[ ,53:104], 1, mean)
```

Calculate Fold Change 

$$
log_2[(T)/N]
$$

T = tumor 
N = normal 

```{r}
e_mean_tumor <- 2^mean_tumor
e_mean_normal <- 2^mean_normal
FoldChange <- e_mean_tumor / e_mean_normal

# Scale Fold Change 
log2_FoldChange <- log2(FoldChange)
```

Calculate significane of the fold change (paired wilcox-test)

```{r}
# calculate p-value with apply function 

p_value <- apply(PRAD_tumor_normal, 1, function(x) wilcox.test(x[1:52], 
                                                         x[53:104], 
                                                         paired = TRUE,
                                                         exact = FALSE)$p.value)

# we will continue working with this p-values

# Scale p-values 
log10_p_value <- log10(p_value)
```


```{r}
# calculate p-value with for loop; stored in data frame

p_value_2 <- data.frame(NA_col = rep(NA, 8159))



for(index in 1:nrow(PRAD_tumor_normal)) {
  p_value_2[index, ] <- wilcox.test(as.matrix(PRAD_tumor_normal[index,1:52]), 
                                    as.matrix(PRAD_tumor_normal[index,53:104]),
                                    paired = TRUE,
                                    exact = FALSE)$p.value
  }

colnames(p_value_2) <- c("p-value")
```


```{r}
# calculate p-value with for loop; stored in matrix

p_value_3 <- matrix(nrow = 8159, ncol = 1)

for(index in 1:nrow(PRAD_tumor_normal)) {
  p_value_3[index, ] <- wilcox.test(as.numeric(PRAD_tumor_normal[index,1:52]), 
                                    as.numeric(PRAD_tumor_normal[index,53:104]),
                                    paired = TRUE,
                                    exact = FALSE)$p.value
  }
```

Plot not corrected p-values 

```{r}
# not correced p-values 
hist(p_value, breaks = 20)
```

correct p-values and plot them again 

```{r}
# correct p-value via Bonferroni ("bonferroni") / Benjamini-Hichberg ("BH")
corrected_p_value <- p.adjust(p_value, method = "bonferroni")

# plot correced p-values 
hist(corrected_p_value, breaks = 20)

# Scale corrected p-values 
log10_corrected_p_value <- log10(corrected_p_value)

```
Create volcano plot with not correced p-values 

```{r}
dat <- data.frame(log2_FoldChange, log10_p_value)

# add a column of NAs
dat$diffexpressed <- "NO"
# if log2Foldchange > 1 and pvalue < 0.05, set as "UP" 
dat$diffexpressed[dat$log2_FoldChange > log2(2) & dat$log10_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < 1 and pvalue < 0.05, set as "DOWN"
dat$diffexpressed[dat$log2_FoldChange < log2(0.5) & dat$log10_p_value < log10(0.05)] <- "DOWN"

# Change point appearance 
color <- c("blue", "red", "black")
names(color) <- c("DOWN", "UP", "NO")

library(ggplot2)

ggplot(data = dat, mapping = aes(x=log2_FoldChange, 
                                 y= -log10_p_value, 
                                 col = diffexpressed))+
  geom_point()+
  scale_color_manual(values = color)+
  geom_vline(xintercept=c(log2(0.5), log2(2)), linetype="dashed", col="black")+
  geom_hline(yintercept=-log10(0.05), linetype="dashed", col="black")+
  scale_x_continuous(breaks = c(seq(-7, 7, 2)),
                     limits = c(-7, 7))+
  labs(title = "Volcano plot")
```

Create volcano plot with corrected p-values 

```{r}
dat_2 <- data.frame(log2_FoldChange, log10_corrected_p_value)

# add a column of NAs
dat_2$diffexpressed <- "NO"
# if log2Foldchange > 1 and pvalue < 0.05, set as "UP" 
dat_2$diffexpressed[dat_2$log2_FoldChange > log2(2) & dat_2$log10_corrected_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < -1 and pvalue < 0.05, set as "DOWN"
dat_2$diffexpressed[dat_2$log2_FoldChange < log2(0.5) & dat_2$log10_corrected_p_value < log10(0.05)] <- "DOWN"

# Change point appearance 
color <- c("blue", "red", "black")
names(color) <- c("DOWN", "UP", "NO")

library(ggplot2)

ggplot(data = dat_2, mapping = aes(x= log2_FoldChange, 
                                   y= -log10_corrected_p_value, 
                                   col= diffexpressed))+
  geom_point()+
  scale_color_manual(values = color)+
  geom_vline(xintercept=c(log2(0.5), log2(2)), linetype="dashed", col="darkgreen")+
  geom_hline(yintercept=-log10(0.05), linetype="dashed", col="darkgreen")+
  scale_x_continuous(breaks = c(seq(-7, 7, 2)),
                     limits = c(-7, 7))+
  labs(title = "Volcano plot")
```

```{r}
# 998 downregulated genes
a <- which(dat_2$diffexpressed == "DOWN")
length(a)

# 347 upregulated genes 
b <- which(dat_2$diffexpressed == "UP")
length(b)
```


Create volcano plot with genes of one pathway 

```{r}
# Create Function to do volcano plot for all pathways


dat_all <- data.frame(log2_FoldChange, log10_corrected_p_value)

# add a column of NAs
dat_all$diffexpressed <- "NO"
# if log2Foldchange > 1 and pvalue < 0.05, set as "UP" 
dat_all$diffexpressed[dat_all$log2_FoldChange > log2(2) & dat_all$log10_corrected_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < -1 and pvalue < 0.05, set as "DOWN"
dat_all$diffexpressed[dat_all$log2_FoldChange < log2(0.5) & dat_all$log10_corrected_p_value < log10(0.05)] <- "DOWN"

# character of all genes 
all_genes <- rownames(PRAD_tumor_normal)
class(all_genes)
# character of all genes in pathway
total_pathways_ensID <- readRDS("~/GitHub/2022-topic-02-team-04/data/total_pathways_ensID.rds")

pathway_genes <- total_pathways_ensID$KEGG_PROSTATE_CANCER
class(pathway_genes)
# add column: TRUE if gene is in pathway + FALSE if it is not
dat_all$pathway_genes <- !is.na(match(all_genes, pathway_genes))

# filter data frame to get TRUE genes in pathway
library(dplyr)
pathway_df <- dat_all %>% 
  filter(pathway_genes == "TRUE")


# Change appearance 
color <- c("blue", "green", "black")
names(color) <- c("DOWN", "UP", "NO")

library(ggplot2)

ggplot(data = dat_all, mapping = aes(x = log2_FoldChange, 
                                     y = -log10_corrected_p_value))+
  geom_point(alpha=0.4,
             size=0.8,
             aes(col = diffexpressed))+
  geom_point(data = pathway_df, mapping = aes(x = log2_FoldChange,
                                              y = -log10_corrected_p_value,
                                              shape = pathway_genes),
             color = "red",
             size = 3,
             alpha = 4)+
  scale_color_manual(values = color)+
  geom_vline(xintercept = c(log2(0.5), log2(2)), linetype = "dashed", col = "darkgreen")+
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", col = "darkgreen")+
  scale_x_continuous(breaks = c(seq(-7, 7, 2)),
                     limits = c(-7, 7))+
  labs(
    title = "Volcano plot",
    subtitle = "Genes in pathway",
    x = "log2 Fold Change",
    y = "- log10 corrected p-value",
    colour = "differentially\nexpressed genes",
    shape = "genes in pathway")
```

Create function to do volcano plot for all data sets 

```{r}
volcano_pathway = function(pathway){
  
  dat_all <- data.frame(log2_FoldChange, log10_corrected_p_value)

  # add a column of NAs
  dat_all$diffexpressed <- "NO"
  # if log2Foldchange > 1 and pvalue < 0.05, set as "UP" 
  dat_all$diffexpressed[dat_all$log2_FoldChange > log2(2) & 
                          dat_all$log10_corrected_p_value <log10(0.05)] <- "UP"
  # if log2Foldchange < -1 and pvalue < 0.05, set as "DOWN"
  dat_all$diffexpressed[dat_all$log2_FoldChange < log2(0.5) & 
                          dat_all$log10_corrected_p_value < log10(0.05)] <- "DOWN"
  
  # character of all genes 
  all_genes <- rownames(PRAD_tumor_normal)
  # character of all genes in pathway
  total_pathways_ensID <- readRDS("~/GitHub/2022-topic-02-team-04/data/total_pathways_ensID.rds")
  
  pathway_genes <- total_pathways_ensID[[pathway]]
  # add column: TRUE if gene is in pathway + FALSE if it is not
  dat_all$pathway_genes <- !is.na(match(all_genes, pathway_genes))
  
  # filter data frame to get TRUE genes in pathway
  library(dplyr)
  pathway_df <- dat_all %>% 
    filter(pathway_genes == "TRUE")
  
  
  # Change appearance 
  color <- c("blue", "green", "black")
  names(color) <- c("DOWN", "UP", "NO")
  
  library(ggplot2)
  
  p <- ggplot(data = dat_all, mapping = aes(x = log2_FoldChange, 
                                       y = -log10_corrected_p_value))+
    geom_point(alpha=0.4,
               size=0.8,
               aes(col = diffexpressed))+
    geom_point(data = pathway_df, mapping = aes(x = log2_FoldChange,
                                                y = -log10_corrected_p_value,
                                                shape = pathway_genes),
               color = "red",
               size = 3,
               alpha = 4)+
    scale_color_manual(values = color)+
    geom_vline(xintercept = c(log2(0.5), log2(2)), linetype = "dashed", col = "darkgreen")+
    geom_hline(yintercept = -log10(0.05), linetype = "dashed", col = "darkgreen")+
    scale_x_continuous(breaks = c(seq(-7, 7, 2)),
                       limits = c(-7, 7))+
    labs(
      title = "Volcano plot",
      subtitle = pathway,
      x = "log2 Fold Change",
      y = "- log10 corrected p-value",
      colour = "differentially expressed genes",
      shape = pathway)
  
  
  
  filename = paste("volcanoplot_", pathway, ".pdf", sep = "")
  pdf(file=filename)
  print(p)
  dev.off()
}


```

```{r}
pathways <- names(total_pathways_ensID)

lapply(pathways, volcano_pathway)
```
