==> prad_normal_vs_tumor/05_size_ratio_pw.Rmd <==
---
title: "Size_ratio_diffexpressed_pathways"
author: "Laura Lange"
date: '2022-07-05'
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load data

```{r}
setwd('..')
tcga_tumor_norm <- readRDS("data/PRAD_TvsN.rds")
tcga_tumor_norm <- readRDS("~/GitHub/2022-topic-02-team-04/data/PRAD_TvsN.rds")

# Fuse data frame PRAD normal with PRAD tumor
PRAD_tumor_normal <- cbind(tcga_tumor_norm[["tumor"]], tcga_tumor_norm[["normal"]])
```

Calculate mean for each gene 

```{r}
mean_tumor <- apply(PRAD_tumor_normal[ ,1:52], 1, mean)

mean_normal <- apply(PRAD_tumor_normal[ ,53:104], 1, mean)
```

Calculate Fold Change 

$$
log_2[(T)/N]
$$

T = tumor 
N = normal 

```{r}
e_mean_tumor <- 2^mean_tumor
e_mean_normal <- 2^mean_normal
FoldChange <- e_mean_tumor / e_mean_normal

# Scale Fold Change 
log2_FoldChange <- log2(FoldChange)
```

Calculate significane of the fold change (paired wilcox-test)

```{r}
# calculate p-value with apply function 

p_value <- apply(PRAD_tumor_normal, 1, function(x) wilcox.test(x[1:52], 
                                                         x[53:104], 
                                                         paired = TRUE,
                                                         exact = FALSE)$p.value)

# Scale p-values 
log10_p_value <- log10(p_value)
```

correct p-values and plot them again 

```{r}
# correct p-value via Bonferroni ("bonferroni") / Benjamini-Hichberg ("BH")
corrected_p_value <- p.adjust(p_value, method = "bonferroni")

# Scale corrected p-values 
log10_corrected_p_value <- log10(corrected_p_value)
```

Get up and downregulated pathways 

```{r}
genes_diff <- data.frame(log2_FoldChange, log10_corrected_p_value)

# add a column of NAs
genes_diff$diffexpressed <- "NO"
# if log2Foldchange > 1 and pvalue < 0.05, set as "UP" 
genes_diff$diffexpressed[genes_diff$log2_FoldChange > log2(2) & genes_diff$log10_corrected_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < -1 and pvalue < 0.05, set as "DOWN"
genes_diff$diffexpressed[genes_diff$log2_FoldChange < log2(0.5) & genes_diff$log10_corrected_p_value < log10(0.05)] <- "DOWN"
```

Load pathways 

```{r}
total_pathways_ensID <- readRDS("~/GitHub/2022-topic-02-team-04/data/total_pathways_ensID.rds")
```

Get size ratio (first try with one example)

```{r}
# number of genes in pathway (not corrected)
# num_pw_genes <- length(pathway_genes)


# gens in pathway (corrected)
pathway_genes <- total_pathways_ensID$PGF_UP.V1_UP

all_pathway_ensID <- unlist(total_pathways_ensID)
pw_genes <- !is.na(match(pathway_genes, all_pathway_ensID))
# number of genes in pathway (corrected)
pw_num <- length(which(pw_genes == TRUE))


# get all diffexpressed genes
genes_NO <- which(genes_diff$diffexpressed == "NO") 
genes_up_down <- genes_diff[-genes_NO, ]
name_gene_diff <- rownames(genes_up_down)
# TRUE if gene is in pathway + FALSE if it is not
genes_pw_diff <- !is.na(match(name_gene_diff, pathway_genes))

# number of diffexpressed genes in pathway
num_diff_genes <- length(which(genes_pw_diff == TRUE))


# calculate ratio
size_ratio <- num_diff_genes/pw_num
```

Create function 

```{r}
size_ratio_pathway = function(pathway){
  
  # gens in pathway
  pathway_genes <- pathway

  all_pathway_ensID <- unlist(total_pathways_ensID)
  pw_genes <- !is.na(match(pathway_genes, all_pathway_ensID))
  # number of genes in pathway (corrected)
  pw_num <- length(which(pw_genes == TRUE))
  # gens in pathway
  
  
  # get all diffexpressed genes
  genes_NO <- which(genes_diff$diffexpressed == "NO") 
  genes_up_down <- genes_diff[-genes_NO, ]
  name_gene_diff <- rownames(genes_up_down)
  # TRUE if gene is in pathway + FALSE if it is not
  genes_pw_diff <- !is.na(match(name_gene_diff, pathway_genes))
  
  # number of diffexpressed genes in pathway
  num_diff_genes <- length(which(genes_pw_diff == TRUE))


  # calculate ratio
  size_ratio <- num_diff_genes/pw_num
  
  return(size_ratio)
}
```

Crate list with unregulated pathways

upregulated pathways 
"Replication_Telomerase"                          
"HALLMARK_ANDROGEN_RESPONSE"                      
"HALLMARK_E2F_TARGETS"                            
"HALLMARK_G2M_CHECKPOINT"                         
"HALLMARK_MITOTIC_SPINDLE"                        
"HALLMARK_MYC_TARGETS_V1"                         
"HALLMARK_MYC_TARGETS_V2"                         
"KEGG_AMINO_SUGAR_AND_NUCLEOTIDE_SUGAR_METABOLISM"
"KEGG_CELL_CYCLE"                                 
"KEGG_ENDOMETRIAL_CANCER"                         
"KEGG_NOTCH_SIGNALING_PATHWAY"

downregulated pathways 
"HALLMARK_MYOGENESIS"                                
"KEGG_DORSO_VENTRAL_AXIS_FORMATION"                  
"KEGG_GLYCOSAMINOGLYCAN_BIOSYNTHESIS_KERATAN_SULFATE"
"KEGG_GLYCOSPHINGOLIPID_BIOSYNTHESIS_GANGLIO_SERIES" 
"BCAT_BILD_ET_AL_UP"

```{r}
up_expressed_pw <- c( "Replication_Telomerase",                    
                      "HALLMARK_ANDROGEN_RESPONSE",                      
                      "HALLMARK_E2F_TARGETS",                           
                      "HALLMARK_G2M_CHECKPOINT",                      
                      "HALLMARK_MITOTIC_SPINDLE",                       
                      "HALLMARK_MYC_TARGETS_V1",                        
                      "HALLMARK_MYC_TARGETS_V2",
                      "KEGG_AMINO_SUGAR_AND_NUCLEOTIDE_SUGAR_METABOLISM",
                      "KEGG_CELL_CYCLE",                                
                      "KEGG_ENDOMETRIAL_CANCER",                  
                      "KEGG_NOTCH_SIGNALING_PATHWAY")

down_expressed_pw <- c("HALLMARK_MYOGENESIS",                  
                       "KEGG_DORSO_VENTRAL_AXIS_FORMATION",
                       "KEGG_GLYCOSAMINOGLYCAN_BIOSYNTHESIS_KERATAN_SULFATE",
                       "KEGG_GLYCOSPHINGOLIPID_BIOSYNTHESIS_GANGLIO_SERIES", 
                       "BCAT_BILD_ET_AL_UP")



up_pathways_ensID <- total_pathways_ensID[up_expressed_pw]

down_pathways_ensID <- total_pathways_ensID[down_expressed_pw]
```


```{r}

percentage_up <- lapply(up_pathways_ensID, function(x) size_ratio_pathway(pathway = x))

percentage_up <- unlist(percentage_up)

percentage_up <- sort(percentage_up, decreasing = TRUE)

percentage_up
```

```{r}
df_percentage_up <- as.data.frame(percentage_up)
df_percentage_up$pathway_names <- rownames(df_percentage_up)


ggplot(data = df_percentage_up, aes(x = percentage_up, 
                                    y = reorder(pathway_names, percentage_up)))+
  geom_point()+
  labs(title = "Upregulated pathways",
       x = "percentage of diffexpressed genes", 
       y = "data set names")
```


```{r}

percentage_down <- lapply(down_pathways_ensID, function(y) size_ratio_pathway(pathway = y))

percentage_down <- unlist(percentage_down)

percentage_down <- sort(percentage_down, decreasing = TRUE)

percentage_down
```

```{r}
df_percentage_down <- as.data.frame(percentage_down)
df_percentage_down$pathway_names <- rownames(df_percentage_down)


ggplot(data = df_percentage_down, aes(x = percentage_down, 
                                      y = reorder(pathway_names, percentage_down)))+
  geom_point()+
  labs(title = "Downregulated pathways",
       x = "percentage of diffexpressed genes", 
       y = "data set names")
```
