==> prad_normal_vs_tumor/06.1_differentially expressed genes list.Rmd <==
---
title: "Differentially expressed gene list"
date: '2022-06-18'
author: "Lottida Phondeth"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

First load the data
```{r}
setwd('..')
tcga_tumor_norm <- readRDS("data/PRAD_TvsN.rds")

```

To visually grasp where the differentially expressed genes are coming from, the code for the volcano plot was applied here again 
Create a data frame that contains the gene expression values for tumor and normal cell data
```{r}
PRAD_tumor_normal <- cbind(tcga_tumor_norm[["tumor"]], tcga_tumor_norm[["normal"]])

```

calculate the mean and the FC
```{r}
mean_t = apply(PRAD_tumor_normal[,1:52], 1, mean)
mean_n = apply(PRAD_tumor_normal[53:104], 1, mean)
e_mean_tumor <- 2^mean_t
e_mean_normal <- 2^mean_n
FoldChange <- e_mean_tumor / e_mean_normal
log2_FoldChange <- log2(FoldChange)
```

The wilcox test is used to determine the significance of our FC
```{r}
p_value <- apply(PRAD_tumor_normal, 1, function(x) wilcox.test(x[1:52], 
                                                         x[53:104], 
                                                         paired = TRUE,
                                                         exact = FALSE)$p.value)
log10_p_value <- log10(p_value)
```

Plot not corrected p-values
```{r}
hist(p_value, breaks = 20)
```

correct the p-values for multiple testing
```{r}
corrected_p_value <- p.adjust(p_value, method = "bonferroni")
hist(corrected_p_value, breaks = 20)
log10_corrected_p_value <- log10(corrected_p_value)
```
create a volcano plot with not corrected p-values
```{r}
dat <- data.frame(log2_FoldChange, log10_p_value)

# add a column of NAs
dat$diffexpressed <- "NO"
# if log2Foldchange > 0.6 and pvalue < 0.05, set as "UP" 
dat$diffexpressed[dat$log2_FoldChange > log2(2) & dat$log10_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < -0.6 and pvalue < 0.05, set as "DOWN"
dat$diffexpressed[dat$log2_FoldChange < log2(0.5) & dat$log10_p_value < log10(0.05)] <- "DOWN"

# Change point appearance 
color <- c("blue", "red", "black")
names(color) <- c("DOWN", "UP", "NO")

library(ggplot2)

ggplot(data = dat, mapping = aes(x=log2_FoldChange, 
                                 y= -log10_p_value, 
                                 col = diffexpressed))+
  geom_point()+
  scale_color_manual(values = color)+
  geom_vline(xintercept=c(log2(0.5), log2(2)), linetype="dashed", col="black")+
  geom_hline(yintercept=-log10(0.05), linetype="dashed", col="black")+
  scale_x_continuous(breaks = c(seq(-7, 7, 2)),
                     limits = c(-7, 7))+
  labs(title = "Volcano plot")
```

plot volcano plot with corrected p-values
```{r}
dat_corrected <- data.frame(log2_FoldChange, log10_corrected_p_value)

# add a column of NAs
dat_corrected$diffexpressed <- "NO"
# if log2Foldchange > 0.6 and pvalue < 0.05, set as "UP" 
dat_corrected$diffexpressed[dat_corrected$log2_FoldChange > log2(2) & dat_corrected$log10_corrected_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < -0.6 and pvalue < 0.05, set as "DOWN"
dat_corrected$diffexpressed[dat_corrected$log2_FoldChange < log2(0.5) & dat_corrected$log10_corrected_p_value < log10(0.05)] <- "DOWN"

# Change point appearance 
color <- c("blue", "red", "black")
names(color) <- c("DOWN", "UP", "NO")

library(ggplot2)

ggplot(data = dat_corrected, mapping = aes(x= log2_FoldChange, 
                                   y= -log10_corrected_p_value, 
                                   col= diffexpressed))+
  geom_point()+
  scale_color_manual(values = color)+
  geom_vline(xintercept=c(log2(0.5), log2(2)), linetype="dashed", col="darkgreen")+
  geom_hline(yintercept=-log10(0.05), linetype="dashed", col="darkgreen")+
  scale_x_continuous(breaks = c(seq(-7, 7, 2)),
                     limits = c(-7, 7))+
  labs(title = "Volcano plot") +
  geom_hline(yintercept=3.636794, linetype="dotted", col="darkgreen")
```

create a gene list with differentially expressed genes
```{r}
n_df = which(dat_corrected$diffexpressed == "NO")
length(n_df) #6456 genes are not differentially expressed 
dat_df = dat_corrected[-n_df,] #we are left with 1345 genes
```
We are left with 1345 differentially expressed genes
extract the names of these genes
```{r}
dat_df_names = as.list(rownames(dat_df))
```

reduce the PRAD_tumor_normal data frame
```{r}
PRAD_tumor_normal_df = subset(PRAD_tumor_normal,rownames(PRAD_tumor_normal) %in% dat_df_names)
```

generate a Heatmap
```{r}
library(pheatmap)

sample_type <- as.data.frame(c(rep("tumor", 52), rep("normal", 52)))
colnames(sample_type) = "sample"
rownames(sample_type) = colnames(PRAD_tumor_normal_df)

pheatmap(PRAD_tumor_normal_df,
         cluster_rows = T,
         show_rownames = F,
         show_colnames = F,
         fontsize = 8,
         annotation_col = sample_type)
```

What is the correlation between the genes?
```{r}
PRAD_tumor_normal_df_cor = cor(t(PRAD_tumor_normal_df))
```

plot a heatmap of the correlation matrix of the genes
```{r}
library(pheatmap)
pheatmap(PRAD_tumor_normal_df_cor,
         cluster_rows = T,
         show_rownames = F,
         show_colnames = F,
         fontsize = 8)
```

***

GO terms
## Function
```{r}
library(biomaRt)

mart = useEnsembl(dataset = "hsapiens_gene_ensembl", biomart='ensembl')
go.terms <- function(genes, quant){
  require(biomaRt)
  require(GO.db)
  
  ## Retrieve GO terms for input genes
  print("Retrieve GO IDs for input genes")
  list_goterms = getBM(attributes = c("ensembl_gene_id", "go_id"), filters = "ensembl_gene_id", values = genes, mart = mart, useCache = FALSE)
  
  ## only take rows with matched GO term, discard unmatched genes
  clist_goterms = list_goterms[!(is.na(list_goterms$go_id) | list_goterms$go_id==""), ]
  
  ## number of genes matched with GO term
  mgenes <- unique(clist_goterms[1])
  
  # Counting the GO term frequency
  
  ## Number of different GO terms
  goterms <- unique(clist_goterms[2])
  
  ## Set up of df for counting
  goterms_counts <- data.frame(matrix(NA, nrow = nrow(goterms)))
  rownames(goterms_counts) = goterms[,1]
  colnames(goterms_counts) = "counts"
  
  ## Counting of go term frequency
  print("Counting of GO ID frequency")
  for (i in 1:nrow(goterms)){
    goterms_counts[i,] <- length(which(clist_goterms$go_id == goterms[i,]))
  }
  
  
  goids <- rownames(goterms_counts)

  ## Getting the GO terms for every GO ID
  print("Getting the GO terms for every GO ID")
  goTerms_list <- lapply(goids, Term)
  
  go_matrix <- t(rbind(goTerms_list))
  
  rownames(go_matrix) = go_matrix
  
  GO_term <- as.data.frame(rownames(go_matrix))
  
  GO <- cbind(GO_term, goterms_counts)
  
  colnames(GO)[1] <- "GO_term"
  
  print("View GO matrix")
  View(GO)
  print(paste("Genes with matched GO term(s):", nrow(mgenes), "/", length(genes))) 
  
  GO_reduced <- GO[GO$counts >= quantile(GO$counts, quant),]
  
  main <- paste("Frequency of GO terms above", quant, "% quantile")
  
  print("Plot GO frequency")
  plot <- ggplot(data = GO_reduced, aes(x = reorder(GO_term, counts), y = counts)) + 
  geom_bar(stat = "identity") + 
  coord_flip() +
  xlab("GO term") +
  ggtitle(main) +
  theme(axis.title = element_text(size = 15)) +
  theme(axis.text = element_text(size = 15))   
  
  return(plot)
}
```

```{r}
dif_genes <- rownames(dat_df)
go.terms(dif_genes, .999)
```


***

To cut down the dimensions we will choose the first 500 genes of the 1345 differentially expressed genes.
In order to achieve this, we arranged the data frame dat_df according to the log10_corrected_p_value and chose the first 500 genes with the highest pvalue.

```{r}
dat_df_sorted <- dplyr::arrange(dat_df, log10_corrected_p_value) ## sort by pvalue

dat_df_small = dat_df_sorted[1:500,]
```

***

determine the GO terms for the 500 genes

```{r}
dif_500genes <- rownames(dat_df_small)
go.terms(dif_500genes, .995)
```
***

Create lists for each GO term that contain the genes that were assigned to this specific GO term. Lists that contained less than 10 differentially expressed genes were removed.

```{r}
library(biomaRt)

mart = useEnsembl(dataset = "hsapiens_gene_ensembl", biomart='ensembl')

x = getBM(attributes = c("ensembl_gene_id", "go_id"), filters = "ensembl_gene_id", values = dif_500genes, mart = mart, useCache = FALSE)

cx_goterms = x[!(is.na(x$go_id) | x$go_id==""), ]

mxgenes <- unique(cx_goterms[1]) #485genes were matched

goxterms <- unique(cx_goterms[2])#3286 GO terms

GO_terms = cx_goterms[2]

Terms = lapply(GO_terms, Term) ## good practice (do not use "T" as variable)

term = as.data.frame(Terms)

colnames(term) <- "GO_term"

cx_goterms = cbind(cx_goterms, term) # matrix with Ensembl ID, GO ID, and GO term
```

```{r}
go_term_genes = list();go_term_genes = vector('list',length(table(cx_goterms$GO_term)))

names(go_term_genes) = names(table(cx_goterms$GO_term))

for (i in 1:length(go_term_genes)){
  go_term_genes[[i]] = cx_goterms[which(cx_goterms$GO_term == names(go_term_genes)[i]),1]
}

## Remove lists with less than 10 entries
for (i in 1:length(go_term_genes)){
  if (length(go_term_genes[[i]]) < 10){
    go_term_genes[[i]] <- NA
  }
}

go_term_genes <- go_term_genes[!is.na(go_term_genes)]
```
___
***
##### HIER BIN ICH GAR NICHT MEHR SO SICHER WIE NÖTIG DAS IST ._.
TopVar 4000 genes

```{r}
var <- apply(PRAD_tumor_normal, 1, var)
PRAD_tumor_normal_names = as.data.frame(rownames(PRAD_tumor_normal))
colnames(PRAD_tumor_normal_names) = "genes"
PRAD_tumor_normal_names$variance = var
PRAD_tumor_normal_names = PRAD_tumor_normal_names[order(PRAD_tumor_normal_names[,"variance"], decreasing = TRUE),]
topVar4000 = PRAD_tumor_normal_names[1:4000,]
topVar4000_names = topVar4000$genes
```

now reduce the PRAD_tumor_normal matrix accordingly

```{r}
PRAD_tumor_normal_topVar4000 = subset(PRAD_tumor_normal,rownames(PRAD_tumor_normal) %in% topVar4000_names)
```
___
***

have a look at the intersection between the genes describing a certain GO term and the topVar4000 genes
This part is also unneccessary since these genes are the genes from the PRAD_tumor_normal dataframe.

```{r}
library(ggVennDiagram)
library(ggplot2)

inter <- data.frame()
                    
inter <- t(as.data.frame(lapply(go_term_genes, function(goterm){
  length(intersect(goterm, rownames(PRAD_tumor_normal)))
})))

# venn diagram
x <- list("GO_term_genes" = unlist(go_term_genes, use.names = F), PvsN_genes = rownames(PRAD_tumor_normal))

ggVennDiagram(x) +
  ggtitle("Venn diagram of go term genes and focused analysis genes") +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggplot2::scale_fill_gradient(low="lightsalmon",high = "lightblue2")
```
## GO term enrichment analysis

```{r}
View(go_term_genes)
View(PRAD_tumor_normal)
```

perform a GSVA

```{r}
library(GSVA)
```

```{r}
PRAD_tumor_normal_matrix <- as.matrix(PRAD_tumor_normal)

go_activity <- gsva(expr = PRAD_tumor_normal_matrix,
                    gset.idx.list = go_term_genes,
                    kcdf = "Gaussian",
                    min.sz = 3)
```
### PCA/UMAP
```{r}
pca <- Seurat::RunPCA(go_activity)
```

```{r}
sample <- c(rep("tumor", 52), rep("normal", 52))
new.c = as.data.frame(pca@cell.embeddings)
new.coordinates = cbind(new.c, sample)

PCbiplot <- function(coords){
  require(ggplot2)
  require(gridExtra)
  
  a <- ggplot(data = coords, mapping = aes(x = PC_1, y = PC_2, color = sample)) + 
  scale_color_manual(values = c("springgreen4", "firebrick4")) +
  geom_point()
  
  b <- ggplot(data = coords, mapping = aes(x = PC_1, y = PC_3, color = sample)) + 
  scale_color_manual(values = c("springgreen4", "firebrick4")) +
  geom_point()
  
  c <- ggplot(data = coords, mapping = aes(x = PC_1, y = PC_4, color = sample)) + 
  scale_color_manual(values = c("springgreen4", "firebrick4")) +
  geom_point()
  
  d <- ggplot(data = coords, mapping = aes(x = PC_2, y = PC_3, color = sample)) + 
  scale_color_manual(values = c("springgreen4", "firebrick4")) +
  geom_point()
          
  e <- ggplot(data = coords, mapping = aes(x = PC_2, y = PC_4, color = sample)) + 
  scale_color_manual(values = c("springgreen4", "firebrick4")) +
  geom_point()
  
  f <- ggplot(data = coords, mapping = aes(x = PC_3, y = PC_4, color = sample)) + 
  scale_color_manual(values = c("springgreen4", "firebrick4")) +
  geom_point()
  
   ## Export
  #filename = paste("13_1_PRAD_GSVA_PC_biplots.pdf", sep = "")
  #pdf(file = filename, height = 8.27, width = 11.69)
  grid.arrange(a,b,c,d,e,f, ncol = 3, top = "PC biplots PRAD GO enrichment analyis")
  #dev.off()
}

PCbiplot(new.coordinates)
```
UMAP
```{r}
set.seed(123); umap <- uwot::umap(pca@cell.embeddings, metric = "cosine", n_components = 2)

umap_2 <- as.data.frame(umap)

sample <- c(rep("tumor", 52), rep("normal", 52))
umap_2 <- cbind(umap_2, sample)

pal = c("springgreen4", "firebrick3")

## Export
#filename = paste("UMAP_PRAD_combined.pdf", sep = "")
#pdf(file = filename, height = 8.27, width = 11.69)

ggplot(umap_2, aes(x = V1, y = V2, color = sample)) + 
  geom_point(alpha = 1) +
  scale_color_manual(values = pal) +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle(paste("PRAD GO enrichment analysis UMAP"))

#dev.off()
```

Würde es Sinn ergeben die "GO activity" (GSVA scores) anzufärben?
