==> prad_normal_vs_tumor/09_dimension_reduction_tumor.Rmd <==
---
title: "Untitled"
author: "Fabian Strobel"
date: '2022-06-11'
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r}
library(ggplot2)
library(gridExtra)
library(scatterplot3d)
set.seed(123)
```

```{r}
PRAD_TvsN <- readRDS('2022-topic-02-team-04/data/PRAD_TvsN')
```

Combine into on data.frame
```{r}
tumor <- as.matrix(PRAD_TvsN$tumor)
```

```{r}
pca_tumor <- Seurat::RunPCA(tumor)
```

Analysis of variance
```{r}
variance = (pca_tumor@stdev) ##Eigenvalues

barplot(log(variance), main = "log(variance) after pca for tumor samples of PRAD dataset", xlab = "PC_i", ylab = "log(variance)")
```

Scaling on 1 gives proportion of variance in percent
```{r}
prop.var = variance/pca_tumor@misc$total.variance

summary(prop.var)

barplot(prop.var[1:10], main = "Proportion of variance for the first 10 PCs", xlab = "PCs", ylab = "Proportion of variance")
```

```{r}
new.coordinates = as.data.frame(pca_tumor@cell.embeddings)

PCbiplot <- function(coords){
  require(ggplot2)
  require(gridExtra)
  
  a <- ggplot(data = coords, mapping = aes(x = PC_1, y = PC_2)) +
  geom_point()
  
  b <- ggplot(data = coords, mapping = aes(x = PC_1, y = PC_3)) +
  geom_point()
  
  c <- ggplot(data = coords, mapping = aes(x = PC_1, y = PC_4)) +
  geom_point()
  
  d <- ggplot(data = coords, mapping = aes(x = PC_2, y = PC_3)) +
  geom_point()
          
  e <- ggplot(data = coords, mapping = aes(x = PC_2, y = PC_4)) +
  geom_point()
  
  f <- ggplot(data = coords, mapping = aes(x = PC_3, y = PC_4)) +
  geom_point()
  
   ## Export
  filename = paste("PCA_biplots_PRAD_tumor.pdf", sep = "")
  pdf(file = filename, height = 8.27, width = 11.69)
  grid.arrange(a,b,c,d,e,f, ncol = 3, top = "PC biplots for focussed analysis (tumor)")
  dev.off()
}

PCbiplot(new.coordinates)
```

```{r}
# n_neighbors
x <- 2

set.seed(123); umap_tumor <- uwot::umap(pca_tumor@cell.embeddings, metric = "cosine", n_components = 2)

umap_tumor <- as.data.frame(umap_tumor)

## Export
filename = paste("UMAP_PRAD_tumor.pdf", sep = "")
pdf(file = filename, height = 8.27, width = 11.69)

ggplot(umap_tumor, aes(x = V1, y = V2)) + 
  geom_point(alpha = 0.5) +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle(paste("UMAP PRAD tumor with n_neighbors ="))

dev.off()
```

```{r}
set.seed(123)
umap_tumor <- uwot::umap(pca_tumor@cell.embeddings, n_neighbors = 2, n_components = 3)
scatterplot3d(umap_tumor,
              type = "h",
              main = "3D Scatter Plot after UMAP")
```
