==> prad_normal_vs_tumor/12_GSVA.Rmd <==
---
title: "GSVA"
author: "Laura Lange"
date: '2022-07-02'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load data 

```{r}
setwd('..')

tcga_tumor_norm <- readRDS("data/PRAD_TvsN.rds")
total_pathways_ensID <- readRDS("data/total_pathways_ensID.rds")

tcga_tumor_norm <- readRDS("~/GitHub/2022-topic-02-team-04/data/PRAD_TvsN.rds")
total_pathways_ensID <- readRDS("~/GitHub/2022-topic-02-team-04/data/total_pathways_ensID.rds")

PRAD_tumor_normal <- cbind(tcga_tumor_norm[["tumor"]], tcga_tumor_norm[["normal"]])
PRAD_anno <- tcga_tumor_norm[["clinical"]]
```

Geneset Variation Analysis (GSVA) Demo [Hänzelmann et al., 2013]

```{r}
library(GSVA)
library(msigdbr)
library(ggplot2)
library(uwot)
```


```{r}
PRAD_tumor_normal <- as.matrix(PRAD_tumor_normal)

pw_activity <- gsva(expr = PRAD_tumor_normal,
                    gset.idx.list = total_pathways_ensID,
                    kcdf = "Gaussian",
                    min.sz = 3)

saveRDS(pw_activity, file = "PRAD_tumor_normal_pw_activity.rds")
```

```{r}
total <- names(total_pathways_ensID)
pw_a <- rownames(pw_activity)

setdiff(total, pw_a)
```
