==> prad_normal_vs_tumor/13_Heatmap_pathway_activity.Rmd <==
---
title: "Heatmap_pathway_activity"
author: "Laura Lange"
date: '2022-07-05'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load data 

```{r}
PRAD_tumor_normal_pw_activity <- readRDS("~/GitHub/2022-topic-02-team-04/data/PRAD_tumor_normal_pw_activity.rds")
```

Heatmap

```{r}
# install.packages("pheatmap")

library(pheatmap)

# clustering method -> 
# "single" -> minimal distance 
# "average" -> average of all pairwise distances 
# "complete" -> maximal distance
# "ward.D2" -> distance between the two cluster is equal to the increase of the variance of the combined cluster compared to individual clusters 
# "centroid" -> distance between the clusters is given by the distance to the centers of the clusters

# distance ->
# "euclidean" -> (Differenz der x-Koordinaten)^2 - Differenz der y-Koordinaten)^2
# "manhattan" -> (Betrage der x-Koordinaten) - (Betrag der y-Koordinaten)
# "correlation"
                      

pheatmap(PRAD_tumor_normal_pw_activity,
         cluster_rows = F,
         show_rownames = F,
         show_colnames = F,
         clustering_distance_cols = "euclidean",
         clustering_method = "ward.D2",
         annotation_col = sample_annot,
         cutree_cols = 2)
```


```{r}
PRAD_GSVA <- as.data.frame(PRAD_tumor_normal_pw_activity)

## Sample annot
sample_annot <- as.data.frame(c(rep("tumor", 52), rep("normal", 52)))
colnames(sample_annot) <- "sample"
rownames(sample_annot) <- colnames(PRAD_GSVA)

## pathway annot

total_pathways_ensID <- readRDS("data/total_pathways_ensID.rds")
totalpw <- names(total_pathways_ensID)
GSVApw <- rownames(PRAD_GSVA)

setdiff(totalpw, GSVApw)

g <- rep("Given Sets", 46)
h <- rep("MSigDB Hallmarks", 50-1)
k <- rep("KEGG", 186-19)
o <- rep("Oncogenic Sigantures", 189)
t <- rep("Ontology", 84)

pw_annot <- as.data.frame(c(g, h, k, o, t))

colnames(pw_annot) <- "Gene sets"
rownames(pw_annot) <- rownames(PRAD_GSVA)

## Heatmap
pheatmap(PRAD_GSVA,
         cluster_rows = T,
         show_rownames = F,
         show_colnames = F,
         annotation_col = sample_annot,
         annotation_row = pw_annot,
         cutree_cols = 2,
         treeheight_row = 0)


```

```{r}
v <- apply(PRAD_GSVA,1,var)

topVar <- PRAD_GSVA[which(v > 0.225),]

## Heatmap
pheatmap(topVar,
         cluster_rows = T,
         show_rownames = F,
         show_colnames = F,
         annotation_col = sample_annot,
         cutree_cols = 2)
```
