==> prad_normal_vs_tumor/15_up_down_regulated_pw_UMAP_pw_activity.Rmd <==
---
title: "Diffexpressed_pathways"
author: "Laura Lange"
date: '2022-07-04'
output: html_document
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load data

```{r}
setwd('..')
tcga_tumor_norm <- readRDS("data/PRAD_tumor_normal_pw_activity.rds")

PRAD_tumor_normal_pw_activity <- readRDS("~/GitHub/2022-topic-02-team-04/data/PRAD_tumor_normal_pw_activity.rds")

total_pathways_ensID <- readRDS("~/GitHub/2022-topic-02-team-04/data/total_pathways_ensID.rds")
```

Calculate mean for each gene 

```{r}
mean_tumor <- apply(PRAD_tumor_normal_pw_activity[ ,1:52], 1, mean)

mean_normal <- apply(PRAD_tumor_normal_pw_activity[ ,53:104], 1, mean)
```

Calculate Fold Change 

$$
log_2[(T)/N]
$$

T = tumor 
N = normal 

```{r}
e_mean_tumor <- 2^mean_tumor
e_mean_normal <- 2^mean_normal
FoldChange <- e_mean_tumor / e_mean_normal
# Scale Fold Change 
log2_FoldChange <- log2(FoldChange)
```

Calculate significane of the fold change (paired wilcox-test)

```{r}
# calculate p-value with apply function 
p_value <- apply(PRAD_tumor_normal_pw_activity, 1, function(x) wilcox.test(x[1:52], 
                                                                           x[53:104],
                                                                           paired = TRUE,
                                                                           exact = FALSE)$p.value)

# Scale p-values 
log10_p_value <- log10(p_value)
```

correct p-values and plot them again 

```{r}
# correct p-value via Bonferroni ("bonferroni") / Benjamini-Hichberg ("BH")
corrected_p_value <- p.adjust(p_value, method = "bonferroni")

# Scale corrected p-values 
log10_corrected_p_value <- log10(corrected_p_value)

```

Create volcano plot with corrected p-values 

```{r}
pw_df <- data.frame(log2_FoldChange, log10_corrected_p_value)

# add a column of NAs
pw_df$diffexpressed <- "NO"
# if log2Foldchange > 0.4 and pvalue < 0.05, set as "UP" 
pw_df$diffexpressed[pw_df$log2_FoldChange > 0.4 & pw_df$log10_corrected_p_value < log10(0.05)] <- "UP"
# if log2Foldchange < -0.4 and pvalue < 0.05, set as "DOWN"
pw_df$diffexpressed[pw_df$log2_FoldChange < -0.4 & pw_df$log10_corrected_p_value < log10(0.05)] <- "DOWN"
```

create a data frame with differentially expressed pathways

```{r}
NO <- which(pw_df$diffexpressed == "NO")
length(NO) #519 pathways are not differentially expressed 
pw_diffexpressed <- pw_df[-NO,] #we are left with 16 pathways
```
create a data frame with up and downregulated pathways

```{r}
pw_upreg <- pw_diffexpressed[-which(pw_diffexpressed$diffexpressed == "DOWN"), ]

pw_downreg <- pw_diffexpressed[-which(pw_diffexpressed$diffexpressed == "UP"), ]
```

Rank up and downregulated pathways

```{r}
pw_upreg$ranks = rank(-pw_upreg$log2_FoldChange)

pw_downreg$ranks = rank(pw_downreg$log2_FoldChange)
```

Extract pathway size

```{r}
for (index in 1:length(rownames(pw_upreg))) {
  pw_name <- rownames(pw_upreg)[index]
  pw_upreg$pw_size[index] <- length(total_pathways_ensID[[pw_name]])
}

for (index in 1:length(rownames(pw_downreg))) {
  pw_name <- rownames(pw_downreg)[index]
  pw_downreg$pw_size[index] <- length(total_pathways_ensID[[pw_name]])
}
```


```{r}
library(ggrepel)
library(ggplot2)

ggplot(data = pw_upreg, mapping = aes(x = ranks, 
                                      y = log2_FoldChange))+
  geom_point(aes(size = pw_size))+
  geom_text_repel(aes(label = rownames(pw_upreg)),
                  size = 2)+
  labs(title = "Upregulated pathways",
       x = "Ranks",
       y = "corrected log10 p-value")
```

```{r}
ggplot(data = pw_downreg, mapping = aes(x = ranks, 
                                        y = log2_FoldChange))+
  geom_point(aes(size = pw_size))+
  geom_text_repel(aes(label = rownames(pw_downreg)),
                  size = 2,
                  box.padding = 1
                  )+
  labs(title = "Downregulated pathways",
       x = "Ranks",
       y = "corrected log10 p-value")
```

Visualize pathway activity by dotplot

```{r}
pw_upreg$pw_names <- rownames(pw_upreg)

ggplot(data = pw_upreg, aes(x = log2_FoldChange, 
                            y = reorder(pw_names, log2_FoldChange)))+
  geom_point(aes(size = pw_size))+
  labs(title = "Upregulated pathways",
       x = "log2 fold change", 
       y = "data set names")
```

```{r}
pw_downreg$pw_names <- rownames(pw_downreg)

ggplot(data = pw_downreg, aes(x = log2_FoldChange, 
                              y = reorder(pw_names, -log2_FoldChange)))+
  geom_point(aes(size = pw_size))+
  labs(title = "Downregulated pathways",
       x = "log2 fold change", 
       y = "data set names")
```


***

###
Visualizing the pathway activity in UMAP
###

```{r}
library(dplyr)
library(ggplot2)
library(gridExtra)
```

```{r}
## PCA
pca <- Seurat::RunPCA(PRAD_tumor_normal_pw_activity)

## UMAP
set.seed(123); umap <- as.data.frame(uwot::umap(pca@cell.embeddings, n_neighbors = 15, metric = "cosine", n_components = 2))
```

## UPREG

```{r}
## Sort by FC
pw_upreg_sorted <- arrange(pw_upreg, desc(log2_FoldChange))

## Get Pathway names
pw_upreg_genes <- rownames(pw_upreg_sorted)

## Get GSVA scores from
pw_upreg_GSVA <- t(PRAD_tumor_normal_pw_activity[pw_upreg_genes,])

## Combine UMAP and Pathway activity
umap_act <- cbind(umap, pw_upreg_GSVA)
```

```{r}
ggplot(umap_act, aes(x = V1, y = V2)) + 
  geom_point(alpha = 1, aes_string(color = noquote(names(umap_act)[4]))) +
  scale_color_gradient(low = "darkblue", high = "yellow") +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle(paste(names(umap_act[4]))) +
  labs(color = "Pathway\nactivity")
```

```{r}
for(i in 3:13) { 
  nam <- paste("A", i, sep = "")
  
  plots_i <- ggplot(umap_act, aes(x = V1, y = V2)) + 
  geom_point(alpha = 1, aes_string(color = noquote(names(umap_act)[i]))) +
  scale_color_gradient(low = "darkblue", high = "yellow") +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle(paste(names(umap_act[i]))) +
  theme(plot.title = element_text(size = 6)) +
  labs(color = "Pathway\nactivity")
  
    assign(nam, plots_i)
}

grid.arrange(A3, A4, A5, A6, A7, A8, A9, A10, A11, A12, A13,
             ncol = 3, 
             top = paste("UMAP after GSVA with pathway activities"))
```

## Report
```{r}
KEGG_DNA_REPLICATION
KEGG_MISMATCH_REPAIR
KEGG_HOMOLOGOUS_RECOBINATION
HALLMARK_E2F_TARGETS
HALLMARK_MYC_TARGETS_V1
SINGH_KRAS_DEPENDENCY_SIGNATURE
KEGG_PROTEASOME
HALLMARK_G2M_CHECKPOINT
KEGG_NUCLEOTIDE_EXCISION_REPAIR
```

```{r}
pw_plot <- rbind(pw_upreg_sorted[3,], pw_upreg_sorted[5,], pw_upreg_sorted[6,], pw_upreg_sorted[7,], pw_upreg_sorted[8,], pw_upreg_sorted[11,])

## Get Pathway names
pw_plot_genes <- rownames(pw_plot)

## Get GSVA scores from
pw_plot_GSVA <- t(PRAD_tumor_normal_pw_activity[pw_plot_genes,])

## Combine UMAP and Pathway activity
umap_act <- cbind(umap, pw_plot_GSVA)

## Plots
for(i in 3:8) { 
  nam <- paste("A", i, sep = "")
  
  plots_i <- ggplot(umap_act, aes(x = V1, y = V2)) + 
  geom_point(alpha = 1, aes_string(color = noquote(names(umap_act)[i]))) +
  scale_color_gradient(low = "darkblue", high = "yellow") +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle(paste(names(umap_act[i])))  +
  labs(color = "Pathway\nactivity")
  
    assign(nam, plots_i)
}

grid.arrange(A3, A4, A5, A6, A7, A8,
             ncol = 6, 
             top = paste("UMAP after GSVA with pathway activities (upregulated)"))
```

```{r}
A8 <- ggplot(umap_act, aes(x = V1, y = V2)) + 
  geom_point(alpha = 1, aes_string(color = noquote(names(umap_act)[8]))) +
  scale_color_gradient(low = "darkblue", high = "yellow") +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle("KEGG_AMINO_SUGAR_AND\nNUCLEOTIDE_SUGAR_METABOLSIM") +
  labs(color = "Pathway\nactivity")
```


## DOWNREG
```{r}
## Sort by FC
pw_downreg_sorted <- arrange(pw_downreg, ranks)

## Get Pathway names
pw_downreg_genes <- rownames(pw_downreg_sorted)

## Get GSVA scores from
pw_downreg_GSVA <- t(PRAD_tumor_normal_pw_activity[pw_downreg_genes,])

## Combine UMAP and Pathway activity
umap_act <- cbind(umap, pw_downreg_GSVA)
```

```{r}
for(i in 3:ncol(umap_act)) { 
  nam <- paste("A", i, sep = "")
  
  plots_i <- ggplot(umap_act, aes(x = V1, y = V2)) + 
  geom_point(alpha = 1, aes_string(color = noquote(names(umap_act)[i]))) +
  scale_color_gradient(low = "darkblue", high = "yellow") +
  xlab("UMAP1") +
  ylab("UMAP2") +
  ggtitle(paste(names(umap_act[i]))) +
  theme(plot.title = element_text(size = 6)) +
  labs(color = "Pathway\nactivity")
  
    assign(nam, plots_i)
}

grid.arrange(A3, A4, A5, A6, A7,
             ncol = 3, 
             top = paste("UMAP after GSVA with pathway activities"))
```
